+++ example/lib/main.dart
@@ -11,6 +11,7 @@ import 'screens/colors.dart';
 import 'screens/flyouts.dart';
 import 'screens/forms.dart';
 import 'screens/icons.dart';
+import 'screens/info_bars.dart';
 import 'screens/inputs.dart';
 import 'screens/mobile.dart';
 import 'screens/commandbars.dart';
@@ -226,6 +227,10 @@ class _MyHomePageState extends State<MyHomePage> with WindowListener {
             title: const Text('Flyouts'),
           ),
           PaneItem(
+            icon: const Icon(FluentIcons.info),
+            title: const Text('InfoBar'),
+          ),
+          PaneItem(
             icon: Icon(
               appTheme.displayMode == PaneDisplayMode.top
                   ? FluentIcons.more
@@ -264,6 +269,7 @@ class _MyHomePageState extends State<MyHomePage> with WindowListener {
         Mobile(),
         CommandBars(),
         FlyoutShowcase(),
+        InfoBars(),
         Others(),
         Settings(),
       ]),