+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 Date format: DD/MM/YYYY
 
-## [next]
+## [3.x.0]
 
 - Show menu button on automatic minimal display mode ([#350](https://github.com/bdlukaa/fluent_ui/pull/350))
 - **BREAKING** `Map<ShortcutActivator, Intent>?` is now the typed used on `FluentApp.shortcuts` ([#351](https://github.com/bdlukaa/fluent_ui/pull/351))
@@ -11,6 +11,14 @@ Date format: DD/MM/YYYY
 - Added `AutoSuggestBox.form`, that uses `TextFormBox` instead of `TextBox` ([#353](https://github.com/bdlukaa/fluent_ui/pull/353))
 - Do not overflow when text is too long on `Chip` ([#322](https://github.com/bdlukaa/fluent_ui/issues/322))
 - Add RTL support for `Chip`
+- `Card` updates:
+  - Updated card's background colors
+  - **BREAKING** Removed `Card.elevation`
+  - Added `Card.margin`, which is the margin around the card
+- `InfoBar` updates:
+  - Added `InfoBar.isIconVisible`
+  - Updated its design to follow the new Win UI 3 guidelines
+- Updated `Combobox` and `Button` designs
 
 ## [3.12.0] - Flutter 3.0 - [13/05/2022]